fix: refactor helper class for application tests
