+++ tests/CNAB.Application.Test/CNAB.Application.Test.csproj
@@ -22,6 +22,7 @@
     <ProjectReference Include="..\..\src\CNAB.Application\CNAB.Application.csproj" />
     <ProjectReference Include="..\..\src\CNAB.Domain\CNAB.Domain.csproj" />
     <ProjectReference Include="..\..\src\CNAB.Infra.Data\CNAB.Infra.Data.csproj" />
+    <ProjectReference Include="..\CNAB.TestHelpers\CNAB.TestHelpers.csproj" />
   </ItemGroup>
 
   <ItemGroup>
