+++ tests/CNAB.Application.Test/Mappings/UserMappingTest.cs
@@ -1,7 +1,7 @@
 using CNAB.Application.DTOs.Account;
 using CNAB.Application.Mappings;
-using CNAB.Application.Test.Common;
 using CNAB.Domain.Entities.Account;
+using CNAB.TestHelpers.Factories;
 using FluentAssertions;
 using Mapster;
 
@@ -20,10 +20,13 @@ public class UserMappingTest
     [Fact]
     public void UserToUserDto_Should_Map_Correctly()
     {
+        // Arrange
         var user = ServiceTestFactory.CreateUser();
 
+        // Act
         var userDto = user.Adapt<UserDto>(_config);
 
+        // Assert
         userDto.Should().NotBeNull();
         userDto.Email.Should().Be(user.Email);
         userDto.Password.Should().Be(user.Password);
@@ -33,10 +36,13 @@ public class UserMappingTest
     [Fact]
     public void UserDtoToUser_Should_Map_Correctly()
     {
+        // Arrange
         var userDto = ServiceTestFactory.CreateUserDto();
 
+        // Act
         var user = userDto.Adapt<User>(_config);
 
+        // Assert
         user.Should().NotBeNull();
         user.Email.Should().Be(userDto.Email);
         user.Password.Should().Be(userDto.Password);
@@ -46,10 +52,13 @@ public class UserMappingTest
     [Fact]
     public void LoginToLoginDto_Should_Map_Correctly()
     {
+        // Arrange
         var login = ServiceTestFactory.CreateLogin();
 
+        // Act
         var loginDto = login.Adapt<LoginDto>(_config);
 
+        // Assert
         loginDto.Should().NotBeNull();
         loginDto.Email.Should().Be(login.Email);
         loginDto.Password.Should().Be(login.Password);
@@ -58,10 +67,13 @@ public class UserMappingTest
     [Fact]
     public void UserTokenToUserTokenDto_Should_Map_Correctly()
     {
+        // Arrange
         var userToken = ServiceTestFactory.CreateUserToken();
 
+        // Act
         var userTokenDto = userToken.Adapt<UserTokenDto>(_config);
 
+        // Assert
         userTokenDto.Should().NotBeNull();
         userTokenDto.Token.Should().Be(userToken.Token);
         userTokenDto.Authenticated.Should().Be(userToken.Authenticated);
@@ -72,10 +84,13 @@ public class UserMappingTest
     [Fact]
     public void Mapping_NullObject_Should_ReturnNull()
     {
+        // Arrange
         User nullUser = null;
 
+        // Act
         var userDto = nullUser.Adapt<UserDto>(_config);
 
+        // Assert
         userDto.Should().BeNull();
     }
 }
