+++ tests/CNAB.Application.Test/Services/TransactionServiceTest.cs
@@ -1,9 +1,9 @@
 using CNAB.Application.DTOs;
 using CNAB.Application.Services;
-using CNAB.Application.Test.Common;
 using CNAB.Domain.Entities;
 using CNAB.Domain.Entities.enums;
 using CNAB.Domain.Interfaces;
+using CNAB.TestHelpers.Factories;
 using MapsterMapper;
 using Microsoft.Extensions.Logging;
 using Moq;
